requirements.txt
flask==3.0.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
gunicorn==21.2.0
seaborn==0.12.2
matplotlib==3.8.2
python-dotenv==1.0.0